Use reshepe web vitals to track website performance  #293
I want to track metrics on the friends of foxley website.  Reshepe has a free tier so trying
it out.

I tried the react package `@reshepe-web-vitals/react` but it required `react-router-dom`
npm package as well.  I tried adding it as a peer dependency, but that still didn't work.
I didn't want to install `react-router-dom` just for reshepe.

I used this guide for config
https://docs.reshepe.dev/web-vitals/javascript

I found that the toolbar didn't work.  It loaded the toolbar but with empty stats, and no
error in the console.  I tried changing the script from defer to async to not specified, and
still didn't work