Spiking the new SEO componenet #291
I think this is the basic structure, but it's not using the graphql siteMetaData

Initially I thought I could remove the data from siteMetaData, and hardcode it here.
Other components use the siteMetaData so seems like I should reuse that.

It's a bit strange to configure settings in the gatsby config, only to serve them up
via graphql, rather than just share constants.  I'm not sure whether to keep the
current graphql pattern or not